# Improve dequeuing performance