--- a/.travis.yml
+++ b/.travis.yml
@@ -11,10 +11,12 @@ go:
 env:
   - GOFLAGS=''
   - GOFLAGS='-race'
-  - GOFLAGS=''                    WITH_ETCD=true
-  - GOFLAGS='-race'               WITH_ETCD=true
-  - GOFLAGS='--tags pkcs11'       WITH_PKCS11=true
-  - GOFLAGS='-race --tags pkcs11' WITH_PKCS11=true
+  - GOFLAGS='--tags batched_queue'
+  - GOFLAGS='-race --tags batched_queue'
+  - GOFLAGS=''                           WITH_ETCD=true
+  - GOFLAGS='-race'                      WITH_ETCD=true
+  - GOFLAGS='--tags pkcs11'              WITH_PKCS11=true
+  - GOFLAGS='-race --tags pkcs11'        WITH_PKCS11=true
 
 matrix:
   fast_finish: true