These changes significantly increase the performance of sequencing by batching the deletes from the `Unsequenced` table and inserts into the `SequencedLeafData` table. This undoes some of the work done in #603.

This is a breaking change as it requires adding a new field to the `Unsequenced` table. It would be possible to do a similar thing without making this change but then requires some slightly hacky uses of the existing indexes to avoid removing rows with duplicate `LeafIdentityHash`s. That said I think the changes are worth the short term pain as they significantly improve performance and reduce load on the database.

As I have yet to put together a god Prometheus dashboard I don't have any good charts to illustrate the speed up so here is some log output from a unpatched and patched `trillian_log_signer` using a batch size of `1000` while using `ct_hammer` with `-max_parallel_chains 50 -invalid_chance 0` and an empty tree. This shows (at least as far as I can tell) that this isn't re-introducing any of the previous locking problems on the `Unsequenced` and `SequencedLeafData` tables that were constraining sequencing performance. Note these numbers are from my laptop so the boost should actually be a bit better on real machines.

```
# master
5487756799089772688: processed 1000 items in 0.84 seconds (1190.85 qps)
5487756799089772688: processed 1000 items in 0.84 seconds (1192.62 qps)
5487756799089772688: processed 1000 items in 0.95 seconds (1052.98 qps)
5487756799089772688: processed 1000 items in 0.95 seconds (1047.63 qps)
5487756799089772688: processed 1000 items in 1.03 seconds (972.74 qps)
5487756799089772688: processed 1000 items in 1.00 seconds (1001.31 qps)
5487756799089772688: processed 1000 items in 1.05 seconds (949.57 qps)
5487756799089772688: processed 1000 items in 1.13 seconds (887.82 qps)
5487756799089772688: processed 1000 items in 1.22 seconds (817.28 qps)
5487756799089772688: processed 1000 items in 1.06 seconds (944.25 qps)

# patched
5487756799089772688: processed 1000 items in 0.14 seconds (7335.50 qps)
5487756799089772688: processed 1000 items in 0.16 seconds (6397.87 qps)
5487756799089772688: processed 1000 items in 0.14 seconds (6920.55 qps)
5487756799089772688: processed 1000 items in 0.17 seconds (5918.10 qps)
5487756799089772688: processed 1000 items in 0.21 seconds (4784.84 qps)
5487756799089772688: processed 1000 items in 0.11 seconds (9004.58 qps)
5487756799089772688: processed 1000 items in 0.13 seconds (7871.44 qps)
5487756799089772688: processed 1000 items in 0.17 seconds (5987.53 qps)
5487756799089772688: processed 1000 items in 0.12 seconds (8514.92 qps)
5487756799089772688: processed 1000 items in 0.10 seconds (10255.44 qps)
```