@Martin2112 2017-06-28
I tested something similar and it didn't really make much difference on large trees, which is why it remains on one of my branches. This split the queue into logical buckets and tried to add and remove from different buckets using a time-based ring approach.

I think we really need to see numbers for a large tree, of at least several hundred million entries. This is where #603 came from. Also can we see the query plans being used for the deletes?

Rather than adding the new field you can possibly reuse the 'bucket' field. It's not currently being used and is there to avoid breakage if we did want to adopt the split queue approach.