@rolandshoemaker 2017-08-31
Sorry this has taken me so long to get back to, I've been juggling a bunch of stuff!

I've done some tests with a much larger tree than my initial tests (~200 million entries) which show an admittedly smaller performance improvement (between 30% and 40%) while no new leaves are being added to `Unsequenced` and a small decrease (although this is probably statistically irrelevant given the variance) and added the `ANALYZE` output for the modified DELETE (both attached below).

Given that for certain applications (such as CT) trees are unlikely to actually grow to this size (except in certain cases such as the Pilot style archival/research style logs) I think this performance improvement is still beneficial (also note that the performance is still significantly improved even under insert load for smaller trees, let me know if you'd like to see stats for that as well) and makes sense.

Using the existing `Bucket` column is possible but we'd probably want to come up with a better way of generating the insert ID since truncating the existing hash down to 4 bytes (the default size of INTEGER) would drop the collision resistance to 16 bits which doesn't seem viable (I think? My math might be shaky here) given that it is quite possible tjat the size of `Unsequenced` may grow quite large depending on what parameters are used for the signer. Since the migration to add this new column would be _relatively_ simple (although it'd need to be done when the `Unsequenced` table is empty since there is no way to easily add a default value without some back-fill tool) I'd probably go with just adding a new column but ¯\_(ツ)_/¯.

---

Perf tests against a tree containing ~200 million entries (note these tests were done on a different machine to those above):

```
# current master

## no insert load
2764478521680768074: processed 1000 items in 0.80 seconds (1244.43 qps)
2764478521680768074: processed 1000 items in 0.70 seconds (1436.39 qps)
2764478521680768074: processed 1000 items in 0.67 seconds (1495.31 qps)
2764478521680768074: processed 1000 items in 0.68 seconds (1467.36 qps)
2764478521680768074: processed 1000 items in 0.62 seconds (1602.25 qps)
2764478521680768074: processed 1000 items in 0.81 seconds (1229.71 qps)
2764478521680768074: processed 1000 items in 0.92 seconds (1089.21 qps)
2764478521680768074: processed 1000 items in 0.85 seconds (1175.28 qps)
2764478521680768074: processed 1000 items in 0.86 seconds (1163.13 qps)
2764478521680768074: processed 650 items in 0.57 seconds (1142.57 qps)

min 1089 max 1602 avg 

## insert load
2764478521680768074: processed 1000 items in 1.12 seconds (890.91 qps)
2764478521680768074: processed 1000 items in 1.03 seconds (974.92 qps)
2764478521680768074: processed 1000 items in 1.02 seconds (981.63 qps)
2764478521680768074: processed 1000 items in 0.99 seconds (1011.97 qps)
2764478521680768074: processed 1000 items in 0.98 seconds (1020.74 qps)
2764478521680768074: processed 1000 items in 0.95 seconds (1057.50 qps)
2764478521680768074: processed 1000 items in 0.95 seconds (1053.20 qps)
2764478521680768074: processed 1000 items in 0.98 seconds (1015.64 qps)
2764478521680768074: processed 1000 items in 1.15 seconds (868.32 qps)
2764478521680768074: processed 1000 items in 0.96 seconds (1041.35 qps)

min 868 max 1057 avg 991 

# patched

## no insert load
2764478521680768074: processed 1000 items in 0.58 seconds (1713.23 qps)
2764478521680768074: processed 1000 items in 0.48 seconds (2083.28 qps)
2764478521680768074: processed 1000 items in 0.44 seconds (2297.43 qps)
2764478521680768074: processed 1000 items in 0.46 seconds (2157.61 qps)
2764478521680768074: processed 1000 items in 0.48 seconds (2085.26 qps)
2764478521680768074: processed 1000 items in 0.48 seconds (2065.70 qps)
2764478521680768074: processed 1000 items in 0.49 seconds (2048.88 qps)
2764478521680768074: processed 1000 items in 0.45 seconds (2224.29 qps)
2764478521680768074: processed 1000 items in 0.44 seconds (2296.30 qps)
2764478521680768074: processed 1000 items in 0.45 seconds (2222.24 qps)

min 1713 max 2297 avg 2119

## insert load
2764478521680768074: processed 1000 items in 0.77 seconds (1304.07 qps)
2764478521680768074: processed 1000 items in 0.88 seconds (1130.97 qps)
2764478521680768074: processed 1000 items in 0.96 seconds (1040.52 qps)
2764478521680768074: processed 1000 items in 0.94 seconds (1063.56 qps)
2764478521680768074: processed 1000 items in 0.98 seconds (1025.39 qps)
2764478521680768074: processed 1000 items in 1.00 seconds (1004.06 qps)
2764478521680768074: processed 1000 items in 1.10 seconds (912.86 qps)
2764478521680768074: processed 1000 items in 1.13 seconds (881.95 qps)
2764478521680768074: processed 1000 items in 1.23 seconds (812.91 qps)

min 812 max 1304 avg 917
```

ANALYZE DELETE:

```
MariaDB [second]> ANALYZE DELETE FROM Unsequenced WHERE QueueID IN (...);
+------+-------------+-------------+------+---------------+------+---------+------+------+---------+----------+------------+-------------+
| id   | select_type | table       | type | possible_keys | key  | key_len | ref  | rows | r_rows  | filtered | r_filtered | Extra       |
+------+-------------+-------------+------+---------------+------+---------+------+------+---------+----------+------------+-------------+
|    1 | SIMPLE      | Unsequenced | ALL  | QueueID       | NULL | NULL    | NULL | 8010 | 7999.00 |   100.00 |      12.50 | Using where |
+------+-------------+-------------+------+---------------+------+---------+------+------+---------+----------+------------+-------------+
```