@Martin2112 2017-09-01
Thanks for doing the extra stats. Can you give the specs of the machines you tested with?

I'm now thinking about how we should proceed. I tested several similar schemes and several times convinced myself they were better, only to be disappointed later. However, if you think it works better for you then we should probably consider making it available.

I think it would be good if this could be made an option while we collect more data on performance, and it takes a long time to get trees up to the 800M entries or so we did before. Ideally, people would only need to make the schema change if they intend to enable the option to avoid disrupting existing log users.

Does that sound sensible?